TimesTables options view
The options view is looking OK now so it's time to stop tweaking and have a look at the game itself. This view makes use of a whole bunch of different SwiftUI elements: custom toggle buttons (set at a wonky angle), view modifiers for part of the buttons and border styling, a tweaked (more colourful) segmented Picker, and a reusable TimesTablesText view to encapsulate the styling of the text items.